+++ bcs-services/bcs-cluster-manager/internal/actions/cluster/delete.go
@@ -83,7 +83,7 @@ func (da *DeleteAction) deleteCluster() error {
 	return nil
 }
 
-func (da *DeleteAction) setResp(code uint64, msg string) {
+func (da *DeleteAction) setResp(code uint32, msg string) {
 	da.resp.Code = code
 	da.resp.Message = msg
 	da.resp.Result = (code == types.BcsErrClusterManagerSuccess)
